+++ CHANGELOG.md
@@ -1,15 +1,18 @@
 # Changelog
+- [v3.5.1](#v351)
+- [v3.5.0](#v350)
+- [v3.4.1](#v341)
 - [v3.4.0](#v340)
 - [v3.3.0](#v330)
 - [v3.2.1](#v321)
 - [v3.2.0](#v320)
 - [v3.1.3](#v313)
-- [v3.1.2](#v312)
-- [v3.1.1](#v311)
-- [v3.1.0](#v310)
 <details>
   <summary>Older Releases...</summary>
   <ul>
+  <li><a href="#v312">v3.1.2</a></li>
+  <li><a href="#v311">v3.1.1</a></li>
+  <li><a href="#v310">v3.1.0</a></li>
   <li><a href="#v302">v3.0.2</a></li>
   <li><a href="#v301">v3.0.1</a></li>
   <li><a href="#v300">v3.0.0</a></li>
@@ -166,6 +169,43 @@
 </details>
 
 ------
+# v3.5.1 (2023-11-17)
+
+#### :bug: Bug Fix
+
+- fix: correctly pass through offset from AnimatedGridRows/Columns [#1769](https://github.com/airbnb/visx/pull/1769)
+
+### :memo: Documentation
+
+- docs: add Superstardle in the wild [#1768](https://github.com/airbnb/visx/pull/1768)
+  
+#### :trophy: Contributors
+- [henrinormak](https://github.com/henrinormak)
+- [mcioffi](https://github.com/mcioffi)
+
+# v3.5.0 (2023-11-13)
+
+#### :rocket: Enhancements
+
+- new(geo, vendor): add `d3-geo@3` to `visx-vendor`, support `geoPath.digits(n)` [#1767](https://github.com/airbnb/visx/pull/1767)
+
+### :memo: Documentation
+
+- Add to "In the wild" [#1765](https://github.com/airbnb/visx/pull/1765)
+  
+#### :trophy: Contributors
+- [kachkaev](https://github.com/kachkaev)
+- [physicianfyi](https://github.com/physicianfyi)
+
+# v3.4.1 (2023-10-26)
+
+#### :bug: Bug Fix
+
+- visx-axis AxisRenderer.tsx: resolve circular dependency [#1762](https://github.com/airbnb/visx/pull/1762)
+  
+#### :trophy: Contributors
+- [ithrforu](https://github.com/ithrforu)
+
 # v3.4.0 (2023-10-10)
 
 #### :rocket: Enhancements